+++ app/api/routes/tasks.py
@@ -1,6 +1,6 @@
 """This module defines the routes for the tasks API."""
 
-from fastapi import APIRouter, Depends
+from fastapi import APIRouter, Depends, HTTPException, status
 
 from ...database import TasksDatabase, get_db
 from ...models import Task, UpdateTask
@@ -8,7 +8,12 @@ from ...models import Task, UpdateTask
 router = APIRouter()
 
 
-@router.post("", response_model=Task, name="tasks:create_task")
+@router.post(
+    "",
+    response_model=Task,
+    status_code=status.HTTP_201_CREATED,
+    name="tasks:create_task",
+)
 def create_task(task: Task, db: TasksDatabase = Depends(get_db)) -> Task:
     """Create a new task.
 
@@ -22,8 +27,49 @@ def create_task(task: Task, db: TasksDatabase = Depends(get_db)) -> Task:
     return db.create_task(task)
 
 
+@router.get("/find", response_model=list[Task], name="tasks:find_tasks")
+def find_tasks(
+    title: str | None = None,
+    description: str | None = None,
+    db: TasksDatabase = Depends(get_db),
+) -> list[Task]:
+    """Find tasks by substring in their description.
+
+    Args:
+        title: The substring to search for in the task title.
+        description: The substring to search for in the task description.
+        db: The database dependency.
+
+    Returns:
+        A list of tasks that match the given substring.
+    """
+    tasks = db.find_tasks(title, description)
+    if tasks:
+        return tasks
+
+    raise HTTPException(status_code=404, detail="No tasks found")
+
+
+@router.get("/{task_id}", response_model=Task, name="tasks:get_task")
+def get_task(task_id: int, db: TasksDatabase = Depends(get_db)) -> Task:
+    """Retrieve a task by ID.
+
+    Args:
+        task_id: The ID of the task to retrieve.
+        db: The database dependency.
+
+    Returns:
+        The task with the given ID.
+    """
+    task = db.get_task(task_id)
+    if task:
+        return task
+
+    raise HTTPException(status_code=404, detail="Task not found")
+
+
 @router.get("", response_model=list[Task], name="tasks:get_tasks")
-def read_tasks(
+def get_tasks(
     sort_by: str | None = None,
     top_n: int | None = None,
     db: TasksDatabase = Depends(get_db),
@@ -52,6 +98,8 @@ def read_tasks(
             tasks.sort(key=lambda task: task.status, reverse=True)
         case "created_at":
             tasks.sort(key=lambda task: task.created_at)
+        case "updated_at":
+            tasks.sort(key=lambda task: task.updated_at)
         case "priority":
             pass
         case _:
@@ -60,27 +108,12 @@ def read_tasks(
     return tasks
 
 
-@router.get("/find", response_model=list[Task], name="tasks:find_task")
-def find_tasks(
-    subtitle: str | None = None,
-    subdesc: str | None = None,
-    db: TasksDatabase = Depends(get_db),
-) -> list[Task]:
-    """Find tasks by substring in their description.
-
-    Args:
-        subtitle: The substring to search for in the task title.
-        subdesc: The substring to search for in the task description.
-        db: The database dependency.
-
-    Returns:
-        A list of tasks that match the given substring.
-    """
-    tasks = db.find_tasks_by_description(subtitle, subdesc)
-    return tasks
-
-
-@router.put("/{task_id}", response_model=Task, name="tasks:update_task")
+@router.put(
+    "/{task_id}",
+    response_model=Task,
+    status_code=status.HTTP_202_ACCEPTED,
+    name="tasks:update_task",
+)
 def update_task(
     task_id: int, updated_task: UpdateTask, db: TasksDatabase = Depends(get_db)
 ) -> Task:
@@ -94,7 +127,10 @@ def update_task(
         The updated task.
     """
     task = db.update_task(task_id, updated_task)
-    return task
+    if task:
+        return task
+
+    raise HTTPException(status_code=404, detail="Task not found")
 
 
 @router.delete("/{task_id}", name="tasks:delete_task")