+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ## [Unreleased]
 
+### Added
+
+- `/health` route [!2](https://github.com/riZZZhik/todo_app/pull/2)
+- `updated_at` field to task [!2](https://github.com/riZZZhik/todo_app/pull/2)
+- Tests [!2](https://github.com/riZZZhik/todo_app/pull/2)
+- CI for linters and tests [!2](https://github.com/riZZZhik/todo_app/pull/2)
+
+### Fixed
+
+- create_task on empty database exception [!2](https://github.com/riZZZhik/todo_app/pull/2)
+
 ## [1.0.0] - 2024-03-22
 
 ### Added